Update test code
