--- mechtatel-core/src/test/java/com/github/maeda6uiui/mechtatel/PhysicalObjectTest.java
@@ -1,166 +1,0 @@
-package com.github.maeda6uiui.mechtatel;
-
-import com.github.maeda6uiui.mechtatel.core.Mechtatel;
-import com.github.maeda6uiui.mechtatel.core.MttSettings;
-import com.github.maeda6uiui.mechtatel.core.MttWindow;
-import com.github.maeda6uiui.mechtatel.core.PixelFormat;
-import com.github.maeda6uiui.mechtatel.core.camera.FreeCamera;
-import com.github.maeda6uiui.mechtatel.core.input.keyboard.KeyCode;
-import com.github.maeda6uiui.mechtatel.core.physics.PhysicalBox;
-import com.github.maeda6uiui.mechtatel.core.physics.PhysicalMesh;
-import com.github.maeda6uiui.mechtatel.core.physics.PhysicalObject;
-import com.github.maeda6uiui.mechtatel.core.postprocessing.PostProcessingProperties;
-import com.github.maeda6uiui.mechtatel.core.screen.MttScreen;
-import com.github.maeda6uiui.mechtatel.core.screen.ScreenImageType;
-import com.github.maeda6uiui.mechtatel.core.screen.component.MttModel;
-import com.jme3.bullet.objects.PhysicsBody;
-import org.joml.Vector3f;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.io.IOException;
-import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Random;
-
-public class PhysicalObjectTest extends Mechtatel {
-    private static final Logger logger = LoggerFactory.getLogger(PhysicalObjectTest.class);
-
-    public PhysicalObjectTest(MttSettings settings) {
-        super(settings);
-        this.run();
-    }
-
-    public static void main(String[] args) {
-        MttSettings
-                .load("./Mechtatel/settings.json")
-                .ifPresentOrElse(
-                        PhysicalObjectTest::new,
-                        () -> logger.error("Failed to load settings")
-                );
-    }
-
-    private MttScreen mainScreen;
-    private MttModel level;
-    private MttModel box;
-    private FreeCamera camera;
-    private List<PhysicalObject> physicalObjects;
-
-    private Random random;
-    private int screenshotCount;
-
-    @Override
-    public void onCreate(MttWindow window) {
-        mainScreen = window.createScreen(
-                new MttScreen.MttScreenCreateInfo()
-                        .setUseShadowMapping(true)
-                        .setPostProcessingNaborNames(Arrays.asList("spotlight", "fog"))
-        );
-
-        mainScreen.getShadowMappingSettings().setBiasCoefficient(0.003f);
-
-        PostProcessingProperties ppProperties = mainScreen.getPostProcessingProperties();
-        ppProperties.fog.setStart(10.0f);
-        ppProperties.fog.setEnd(20.0f);
-        ppProperties.spotlightAmbientColor = new Vector3f(0.0f, 0.0f, 0.0f);
-
-        var spotlightR = ppProperties.createSpotlight();
-        spotlightR.setPosition(new Vector3f(10.0f, 10.0f, 10.0f));
-        spotlightR.setDirection(new Vector3f(-1.0f, -1.0f, -1.0f));
-        spotlightR.setDiffuseColor(new Vector3f(1.0f, 0.0f, 0.0f));
-
-        var spotlightG = ppProperties.createSpotlight();
-        spotlightG.setPosition(new Vector3f(10.0f, 10.0f, -10.0f));
-        spotlightG.setDirection(new Vector3f(-1.0f, -1.0f, 1.0f));
-        spotlightG.setDiffuseColor(new Vector3f(0.0f, 1.0f, 0.0f));
-
-        var spotlightB = ppProperties.createSpotlight();
-        spotlightB.setPosition(new Vector3f(-10.0f, 10.0f, 0.0f));
-        spotlightB.setDirection(new Vector3f(1.0f, -1.0f, 0.0f));
-        spotlightB.setDiffuseColor(new Vector3f(0.0f, 0.0f, 1.0f));
-
-        try {
-            level = mainScreen.createModel(Paths.get("./Mechtatel/Standard/Model/Plane/plane.obj"));
-
-            box = mainScreen.createModel(Paths.get("./Mechtatel/Standard/Model/Cube/cube.obj"));
-            box.setVisible(false);
-        } catch (IOException e) {
-            logger.error("Error", e);
-            window.close();
-
-            return;
-        }
-
-        var phLevel = new PhysicalMesh(level, PhysicsBody.massForStatic);
-        phLevel.getBody().setRestitution(0.7f);
-        phLevel.getBody().setFriction(0.5f);
-
-        camera = new FreeCamera(mainScreen.getCamera());
-
-        physicalObjects = new ArrayList<>();
-
-        random = new Random();
-        screenshotCount = 0;
-    }
-
-    @Override
-    public void onUpdate(MttWindow window) {
-        camera.translate(
-                window.getKeyboardPressingCount(KeyCode.W),
-                window.getKeyboardPressingCount(KeyCode.S),
-                window.getKeyboardPressingCount(KeyCode.A),
-                window.getKeyboardPressingCount(KeyCode.D)
-        );
-        camera.rotate(
-                window.getKeyboardPressingCount(KeyCode.UP),
-                window.getKeyboardPressingCount(KeyCode.DOWN),
-                window.getKeyboardPressingCount(KeyCode.LEFT),
-                window.getKeyboardPressingCount(KeyCode.RIGHT)
-        );
-
-        float x = random.nextFloat();
-        float z = random.nextFloat();
-        int signX = (random.nextInt() % 2 == 0) ? 1 : -1;
-        int signZ = (random.nextInt() % 2 == 0) ? 1 : -1;
-        x *= signX;
-        z *= signZ;
-
-        if (window.getKeyboardPressingCount(KeyCode.KEY_1) == 1) {
-            var dupBox = mainScreen.duplicateModel(box);
-            dupBox.rescale(new Vector3f(0.5f, 0.5f, 0.5f));
-
-            var phBox = new PhysicalBox(0.5f, 1.0f);
-            phBox.getBody().setRestitution(0.7f);
-            phBox.getBody().setFriction(0.5f);
-            phBox.setLocation(new Vector3f(x, 10.0f, z));
-            phBox.addComponent(dupBox);
-            physicalObjects.add(phBox);
-        }
-
-        if (window.getKeyboardPressingCount(KeyCode.C) == 1) {
-            physicalObjects.forEach(PhysicalObject::cleanup);
-            physicalObjects.clear();
-        }
-        physicalObjects.forEach(PhysicalObject::syncComponents);
-
-        mainScreen.draw();
-        window.present(mainScreen);
-
-        if (window.getKeyboardPressingCount(KeyCode.F1) == 1) {
-            try {
-                mainScreen.save(
-                        ScreenImageType.COLOR,
-                        PixelFormat.BGRA,
-                        Paths.get(String.format("screenshot_%d.png", screenshotCount))
-                );
-                screenshotCount++;
-            } catch (
-                    IOException e) {
-                logger.error("Error", e);
-                window.close();
-            }
-        }
-    }
-}
